move key_library.py from posydon/binary_evol/DT to posydon/utils
