+++ posydon/binary_evol/track_match.py
@@ -22,7 +22,7 @@ import scipy
 from scipy.optimize import minimize, root
 
 import posydon.utils.constants as const
-from posydon.binary_evol.DT.key_library import (
+from posydon.utils.key_library import (
     DEFAULT_FINAL_KEYS,
     DEFAULT_PROFILE_KEYS,
     DEFAULT_TRANSLATED_KEYS,
